+++ code/packages/python/ir-to-wasm-compiler/tests/test_ir_to_wasm_compiler.py
@@ -202,7 +202,7 @@ def test_compile_syscall_write_uses_wasi_fd_write() -> None:
     program.add_instruction(
         IrInstruction(IrOp.LOAD_IMM, [IrRegister(4), IrImmediate(65)], id=gen.next())
     )
-    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(1)], id=gen.next()))
+    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(1), IrRegister(4)], id=gen.next()))
     program.add_instruction(IrInstruction(IrOp.HALT, [], id=gen.next()))
 
     module = IrToWasmCompiler().compile(
@@ -221,7 +221,7 @@ def test_compile_syscall_read_uses_wasi_fd_read() -> None:
     gen = IDGenerator()
     program = IrProgram(entry_label="_start")
     program.add_instruction(IrInstruction(IrOp.LABEL, [IrLabel("_start")], id=-1))
-    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(2)], id=gen.next()))
+    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(2), IrRegister(4)], id=gen.next()))
     program.add_instruction(
         IrInstruction(IrOp.ADD_IMM, [IrRegister(1), IrRegister(4), IrImmediate(0)], id=gen.next())
     )
@@ -244,7 +244,7 @@ def test_compile_syscall_exit_uses_wasi_proc_exit() -> None:
     program.add_instruction(
         IrInstruction(IrOp.LOAD_IMM, [IrRegister(4), IrImmediate(7)], id=gen.next())
     )
-    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(10)], id=gen.next()))
+    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(10), IrRegister(4)], id=gen.next()))
 
     module = IrToWasmCompiler().compile(
         program,
@@ -260,7 +260,7 @@ def test_compile_syscall_exit_uses_wasi_proc_exit() -> None:
 def test_unsupported_syscall_raises() -> None:
     program = IrProgram(entry_label="_start")
     program.add_instruction(IrInstruction(IrOp.LABEL, [IrLabel("_start")], id=-1))
-    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(99)], id=IDGenerator().next()))
+    program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(99), IrRegister(4)], id=IDGenerator().next()))
 
     try:
         IrToWasmCompiler().compile(
@@ -487,7 +487,7 @@ class TestDispatchLoopLowerer:
         program.add_instruction(
             IrInstruction(IrOp.LOAD_IMM, [IrRegister(4), IrImmediate(72)], id=gen.next())  # 'H'
         )
-        program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(1)], id=gen.next()))
+        program.add_instruction(IrInstruction(IrOp.SYSCALL, [IrImmediate(1), IrRegister(4)], id=gen.next()))
         program.add_instruction(IrInstruction(IrOp.HALT, [], id=gen.next()))
 
         output: list[str] = []